web/requirements-dev.txt
pyformat
isort
doc8
sphinx
sphinx-autobuild
sphinx_rtd_theme

pytest
pytest-xdist
pytest-mock

django-extensions
pygraphviz --install-option="--include-path=/usr/include/graphviz" \
           --install-option="--library-path=/usr/lib/graphviz/"

#pyparsing==1.5.7
#pydot

